FIX: Hello world now works
FIX: MEMSIM can now handle 512b bus width
FIX: BOARD.LD kernel sector deconflicted
FIX: QFLEXPRESS now uses different NDUMMY based upon
  context
WIP: SDINFO ...

+++ sim/memsim.h
@@ -68,12 +68,12 @@ public:
 	void	apply(const uchar wb_cyc, const uchar wb_stb,
 				const uchar wb_we,
 			const BUSW wb_addr, const uint32_t *wb_data,
-				const short wb_sel,
+				const uint64_t wb_sel,
 			uchar &o_stall, uchar &o_ack, uint32_t *o_data);
 	void	operator()(const uchar wb_cyc, const uchar wb_stb,
 				const uchar wb_we,
 			const BUSW wb_addr, const uint32_t *wb_data,
-				const short wb_sel,
+				const uint64_t wb_sel,
 			uchar &o_stall, uchar &o_ack, uint32_t *o_data) {
 
 		apply(wb_cyc, wb_stb, wb_we, wb_addr, wb_data, wb_sel,
